✅ simplify test fixture
