+++ SurrealDb.Net.Tests.Fixtures/DatabaseInfo.cs
@@ -1,0 +1,7 @@
+﻿namespace SurrealDb.Net.Tests.Fixtures;
+
+public sealed class DatabaseInfo
+{
+    public string Namespace { get; init; } = string.Empty;
+    public string Database { get; init; } = string.Empty;
+}
