+++ SurrealDb.Net.Tests.Fixtures/DatabaseInfoFaker.cs
@@ -1,0 +1,12 @@
+﻿using Bogus;
+
+namespace SurrealDb.Net.Tests.Fixtures;
+
+internal sealed class DatabaseInfoFaker : Faker<DatabaseInfo>
+{
+    public DatabaseInfoFaker()
+    {
+        RuleFor(o => o.Namespace, f => f.Random.AlphaNumeric(40));
+        RuleFor(o => o.Database, f => f.Random.AlphaNumeric(40));
+    }
+}
