+++ SurrealDb.Net.Tests.Fixtures/FilePathInfo.cs
@@ -1,0 +1,6 @@
+﻿namespace SurrealDb.Net.Tests.Fixtures;
+
+internal sealed class FilePathInfo
+{
+    public string Path { get; init; } = string.Empty;
+}
